fixed InputHandler created Output class
