+++ src/test/java/test/InputValidatorTest.java
@@ -1,0 +1,7 @@
+package test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class InputValidatorTest {
+
+}
